Fix accidental copy into wrong directory
