+++ widgets/recurrencewidget.cpp
@@ -278,7 +278,7 @@ void RecurrenceWidget::languageChange()
   */
 void RecurrenceWidget::clear()
 {
-  set(false, 1, "W", QDateTime::currentDateTime(), _eot, 1);
+  set(false, 1, "W", QDateTime::currentDateTime(), _eot, 1, "KeepNone");
   _id             = -1;
 
   _prevParentId   = -1;
@@ -419,6 +419,16 @@ bool RecurrenceWidget::maxVisible() const
   return _max->isVisible();
 }
 
+QString RecurrenceWidget::style() const
+{
+  if (_none->isChecked())
+    return "KeepNone";
+  if (_one->isChecked())
+    return "KeepOne";
+  if (_all->isChecked())
+    return "KeepAll";
+}
+
 RecurrenceWidget::RecurrencePeriod RecurrenceWidget::minPeriod() const
 {
   return (RecurrencePeriod)_period->id(0);
@@ -434,7 +444,8 @@ bool RecurrenceWidget::modified() const
              << "endDateTime:"   << endDateTime()  << _prevEndDateTime   << "\n"
              << "max:"           << max()          << _prevMax           << "\n"
              << "_parentId:"     << _parentId      << _prevParentId      << "\n"
-             << "_parentType:"   << _parentType    << _prevParentType
+             << "_parentType:"   << _parentType    << _prevParentType    << "\n"
+             << "_style:"        << style()        << _prevStyle
              ;
 
   bool returnVal = (isRecurring()   != _prevRecurring    ||
@@ -444,7 +455,8 @@ bool RecurrenceWidget::modified() const
                     endDateTime()   != _prevEndDateTime  ||
                     max()           != _prevMax          ||
                     _parentId       != _prevParentId     ||
-                    _parentType     != _prevParentType);
+                    _parentType     != _prevParentType   ||
+                    _prevStyle      != style());
 
   return returnVal;
 }
@@ -576,7 +588,8 @@ bool RecurrenceWidget::save(bool externaltxn, RecurrenceChangePolicy cp, QString
                      "  recur_freq=:recur_freq,"
                      "  recur_start=:recur_start,"
                      "  recur_end=:recur_end,"
-                     "  recur_max=:recur_max"
+                     "  recur_max=:recur_max,"
+                     "  recur_style=:recur_style"
                      " WHERE (recur_id=:recurid)"
                      " RETURNING recur_id;");
       recurq.bindValue(":recurid", _id);
@@ -587,12 +600,12 @@ bool RecurrenceWidget::save(bool externaltxn, RecurrenceChangePolicy cp, QString
                      "  recur_parent_id,  recur_parent_type,"
                      "  recur_period,     recur_freq,"
                      "  recur_start,      recur_end,"
-                     "  recur_max"
+                     "  recur_max,        recur_style"
                      ") VALUES ("
                      "  :recur_parent_id, UPPER(:recur_parent_type),"
                      "  :recur_period,    :recur_freq,"
                      "  :recur_start,     :recur_end,"
-                     "  :recur_max"
+                     "  :recur_max,       :recur_style"
                      ") RETURNING recur_id;");
     }
 
@@ -604,6 +617,7 @@ bool RecurrenceWidget::save(bool externaltxn, RecurrenceChangePolicy cp, QString
     if (endDate() < _eot.date())
       recurq.bindValue(":recur_end",       endDateTime());
     recurq.bindValue(":recur_max",         max());
+    recurq.bindValue(":recur_style", style());
     recurq.exec();
     if (recurq.first())
     {
@@ -616,6 +630,7 @@ bool RecurrenceWidget::save(bool externaltxn, RecurrenceChangePolicy cp, QString
       _prevPeriod        = period();
       _prevRecurring     = isRecurring();
       _prevStartDateTime = startDateTime();
+      _prevStyle         = style();
     }
   }
   else // ! isRecurring()
@@ -701,34 +716,35 @@ bool RecurrenceWidget::save(bool externaltxn, RecurrenceChangePolicy cp, QString
   return true;
 }
 
-void RecurrenceWidget::set(bool recurring, int frequency, QString period, QDate startDate, QDate endDate, int max)
+void RecurrenceWidget::set(bool recurring, int frequency, QString period, QDate startDate, QDate endDate, int max, QString style)
 {
   if (DEBUG)
     qDebug() << objectName() << "::set(" << recurring << ", "
              << frequency    << ", "     << period    << ", "
              << startDate    << ", "     << endDate   << ", "
-             << max          << ") entered";
+             << max          << ", "     << style     << ") entered";
   // run from the beginning of the start date to the end of the end date
   QDateTime startDateTime(startDate);
   QDateTime endDateTime(endDate.addDays(1));
   endDateTime = endDateTime.addMSecs(-1);
 
-  set(recurring, frequency, period, startDateTime, endDateTime, max);
+  set(recurring, frequency, period, startDateTime, endDateTime, max, style);
 }
 
-void RecurrenceWidget::set(bool recurring, int frequency, QString period, QDateTime start, QDateTime end, int max)
+void RecurrenceWidget::set(bool recurring, int frequency, QString period, QDateTime start, QDateTime end, int max, QString style)
 {
   if (DEBUG)
     qDebug() << objectName() << "::set(" << recurring << ", "
              << frequency    << ", "     << period    << ", "
              << start        << ", "     << end       << ", "
-             << max          << ") entered";
+             << max          << ", "     << style     << ") entered";
   setRecurring(recurring);
   setPeriod(period);
   setFrequency(frequency);
   setStartDateTime(start);
   setEndDateTime(end);
   setMax(max);
+  setStyle(style);
 
   _prevEndDateTime   = end.isValid() ? end : _eot ;
   _prevFrequency     = frequency;
@@ -736,6 +752,7 @@ void RecurrenceWidget::set(bool recurring, int frequency, QString period, QDateT
   _prevRecurring     = recurring;
   _prevStartDateTime = start;
   _prevMax           = max;
+  _prevStyle         = style;
 }
 
 void RecurrenceWidget::setEndDate(QDate p)
@@ -789,6 +806,19 @@ void RecurrenceWidget::setMaxVisible(bool p)
   _maxLit->setVisible(p);
 }
 
+void RecurrenceWidget::setStyle(QString style)
+{
+  _none->setChecked(false);
+  _one->setChecked(false);
+  _all->setChecked(false);
+  if (style=="KeepNone")
+    _none->setChecked(true);
+  if (style=="KeepOne")
+    _one->setChecked(true);
+  if (style=="KeepAll")
+    _all->setChecked(true);
+}
+
 void RecurrenceWidget::setMinPeriod(RecurrencePeriod min)
 {
   _period->clear();
@@ -828,7 +858,8 @@ bool RecurrenceWidget::setParent(int pid, QString ptype)
         recurq.value("recur_period").toString(),
         recurq.value("recur_start").toDateTime(),
         recurq.value("recur_end").toDateTime(),
-        recurq.value("recur_max").toInt());
+        recurq.value("recur_max").toInt(),
+        recurq.value("recur_style").toString());
     _id             = recurq.value("recur_id").toInt();
     _prevParentId   = _parentId;
     _prevParentType = _parentType;
