+++ widgets/recurrencewidget.ui
@@ -109,6 +109,30 @@
        </widget>
       </item>
       <item row="4" column="0">
+       <widget class="QRadioButton" name="_none">
+        <property name="text">
+         <string>Keep None</string>
+        </property>
+        <property name="checked">
+         <bool>true</bool>
+        </property>
+       </widget>
+      </item>
+      <item row="4" column="1">
+       <widget class="QRadioButton" name="_one">
+        <property name="text">
+         <string>Keep One</string>
+        </property>
+       </widget>
+      </item>
+      <item row="4" column="2">
+       <widget class="QRadioButton" name="_all">
+        <property name="text">
+         <string>Keep All</string>
+        </property>
+       </widget>
+      </item>
+      <item row="5" column="0">
        <spacer name="verticalSpacer">
         <property name="orientation">
          <enum>Qt::Vertical</enum>
